DBA-791
Add "nolog: true" when setting facts involving passwords or calling scripts passing in the password so that the password is not leaked.
Removed password setting within bash scripts and pass as stdin instead as that can't be read by ps or be recorded in logs.
Changed emcli login command to use password passed in through stdin as it's more secure and again, can't be read by ps, etc.