DBA-920
Add default retention policy value of 1512 hours, which is 9 weeks, instead of 0 as this is only used in the logrotate config and using 0 means it will immediately delete the old logs, which is not want we want to do with the old alert and listener logs. NB it is not used for any other logfiles at this time.
Fixed the logrotate functions to use the monthly value instead of the weeks value as the logs are being rotated on a monthly basis.