DBA-993
Remove Lifecycle Status check from Rule Set as notification validation is carried out in the package and monitoring script.
Add check to package and monitoring script to see if the host that the target relates to has a database on it with the Diagnostics pack. If the host has a db with the pack then we can use the native OEM Notification function, otherwise we can't.